Add ability to store designs and animations using SPIFFS
Add ability to store designs and animations created in web UI to board
flash using SPIFFS. Provide StorageManager class for the project that
handles the storage of the designs, animations and the index files. All
files are stored as JSONs. Also performed code refactoring and
enhancements. The web UI has gallery of designs and animations, but
currently the preview is broken.